requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.95.0
anyio==4.13.0
APScheduler==3.11.2
attrs==26.1.0
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cryptography==46.0.7
distro==1.9.0
docstring_parser==0.18.0
Flask==3.1.3
frozenlist==1.8.0
google-auth==2.49.2
google-auth-oauthlib==1.3.1
gspread==6.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.14.0
MarkupSafe==3.0.3
multidict==6.7.1
oauthlib==3.3.1
propcache==0.4.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.0
pydantic_core==2.46.0
PyJWT==2.12.1
python-dotenv==1.2.2
requests==2.33.1
requests-oauthlib==2.0.0
sniffio==1.3.1
twilio==9.10.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
Werkzeug==3.1.8
yarl==1.23.0
